Fixing Activitys Data base Bug fixed

+++ app/src/main/java/com/example/tavar/arrendaapp/MainActivity.java
@@ -52,7 +52,7 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
     }
 
     private void editHouse(){
-        int id = (int) feedCursorAdapter.getLastHouseClicked();
+        int id = feedCursorAdapter.getLastHouseClicked();
 
         Intent intent = new Intent(this, ActivityEdit.class);
